Extracted interface Leashable. Only golems and animals are leashable in vanilla, However other entities are still able to be leashed if forced.
