+++ src/main/java/org/spongepowered/api/entity/living/golem/Golem.java
@@ -26,10 +26,11 @@
 package org.spongepowered.api.entity.living.golem;
 
 import org.spongepowered.api.entity.living.Agent;
+import org.spongepowered.api.entity.living.Leashable;
 
 /**
  * Represents a Golem type mob.
  */
-public interface Golem extends Agent {
+public interface Golem extends Agent, Leashable {
 
 }
